requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.4
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
flake8-html==0.4.3
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
